backend/requirements.txt
fastapi
uvicorn
numpy
python-dotenv
google-genai